// src/app/features/financial-statement-reports/model/financial-sale-reports.ts
export interface FinancialSaleReport {
    providerProductName?: string;
    departmentName?: string;
    mainsectionName?: string;
    subsectionName?: string;
    productCode?: number;
    ean13?: string;
    dun14?: string;
    saleNumber?: number;
    from?: string;
    to?: string;
    isEffectiveSale?: boolean;
  }
  